Scalable algorithm
Pretty much scrapped the number-only guesser (since program takes input
as a string, isa(realpass, 'double') always returns false).  Figured out
a way to have the program cycle through every combination of numbers,
lower case letters, and a handful of symbols.  It's a little clunky, but
it works up to four characters.  Added a counter (not sure if it's
mathematically accurate, but I think it might be).  Interesting factoid,
the last password the program will try is <<<<, which (according to my
possibly inaccurate counter) takes over 18 million guesses, or about 1
min 45 sec of computation.